Fix yo test.
